Extract FakeReportServer
